requirements.txt
requests==2.26.0
python-telegram-bot==13.7
openai==0.8.0
Flask==2.0.2
